feat(syscalls_table): change the type for the field 'number'

+++ syscalls_table.h
@@ -2,7 +2,7 @@
 #define SYSCALLS_TABLE_H
 
 typedef struct {
-    int number;
+    unsigned long long number;
     const char *name;
 } syscall_info_t;
